spilt settings into 4 components
